requirements.txt
certifi==2019.3.9
chardet==3.0.4
idna==2.8
packaging==19.0
pip-review==1.0
pyparsing==2.4.0
python-dotenv==0.10.3
requests==2.21.0
six==1.12.0
tqdm==4.31.1
urllib3==1.24.2
